Refactor codes.

+++ bin/start-spark-workload.sh
@@ -7,10 +7,10 @@ export HADOOP_CONF_DIR=${HADOOP_HOME}/etc/hadoop
 export JAVA_HOME=$4
 QUEUE=$5
 DATA_SIZE=$6
-WORK_DIR=$7
-JAR_DIR="${WORK_DIR}"/data/workloads
-MNIST_DIR="${WORK_DIR}"/data/mnist
-OUTPUT_DIR="${WORK_DIR}"/results/workload-out
+PROJECT_DIR=$7
+WORKLOAD_JAR_DIR="${PROJECT_DIR}"/data/workloads
+MNIST_DIR="${PROJECT_DIR}"/data/mnist
+OUTPUT_DIR="${PROJECT_DIR}"/results/workload-out
 
 EXECUTOR_CORES=2
 NUM_EXECUTORS=2
@@ -20,7 +20,7 @@ DRIVER_MEMORY=2g
 echo "Starting workload with parameters: class ${CLASS}, queue ${QUEUE}, data_size ${DATA_SIZE}"
 case ${CLASS} in
 "rnn")
-    "${SPARK_HOME}"/bin/spark-submit \
+    spark-submit \
         --class com.intel.analytics.bigdl.models.rnn.workload_rnn \
         --master yarn \
         --deploy-mode cluster \
@@ -36,28 +36,11 @@ case ${CLASS} in
         --checkpoint "${OUTPUT_DIR}" \
         -e "${DATA_SIZE}"
 
-     rm "${OUTPUT_DIR}"/*
-     ;;
-
-"autoencoder" | "lenet")
-    "${SPARK_HOME}"/bin/spark-submit \
-        --class com.intel.analytics.bigdl.models."${CLASS}".workload_"${CLASS}" \
-        --master yarn \
-        --deploy-mode cluster \
-        --executor-cores "${EXECUTOR_CORES}" \
-        --num-executors "${NUM_EXECUTORS}" \
-        --executor-memory "${EXECUTOR_MEMORY}" \
-        --driver-memory "${DRIVER_MEMORY}" \
-        --queue "${QUEUE}" \
-        "${JAR_DIR}" \
-        -b 8 \
-        -f "${MNIST_DIR}"/cifar-10/ \
-        -e "${DATA_SIZE}"
-    rm "${OUTPUT_DIR}"/*
+    rm -r "${OUTPUT_DIR:?}"/*
     ;;
 
-"resnet" | "vgg")
-    "${SPARK_HOME}"/bin/spark-submit \
+"autoencoder" | "lenet")
+    spark-submit \
         --class com.intel.analytics.bigdl.models."${CLASS}".workload_"${CLASS}" \
         --master yarn \
         --deploy-mode cluster \
@@ -70,8 +53,25 @@ case ${CLASS} in
         -b 8 \
         -f "${MNIST_DIR}" \
         -e "${DATA_SIZE}"
+    rm -r "${OUTPUT_DIR:?}"/*
+    ;;
 
-    rm "${OUTPUT_DIR}"/*
+"resnet" | "vgg")
+    spark-submit \
+        --class com.intel.analytics.bigdl.models."${CLASS}".workload_"${CLASS}" \
+        --master yarn \
+        --deploy-mode cluster \
+        --executor-cores "${EXECUTOR_CORES}" \
+        --num-executors "${NUM_EXECUTORS}" \
+        --executor-memory "${EXECUTOR_MEMORY}" \
+        --driver-memory "${DRIVER_MEMORY}" \
+        --queue "${QUEUE}" \
+        "${JAR_DIR}" \
+        -b 8 \
+        -f "${MNIST_DIR}"/cifar-10 \
+        -e "${DATA_SIZE}"
+
+    rm -r "${OUTPUT_DIR:?}"/*
     ;;
 
 *)
@@ -83,6 +83,6 @@ case ${CLASS} in
         --num-executors "${NUM_EXECUTORS}" \
         --executor-memory "${EXECUTOR_MEMORY}" \
         --driver-memory "${DRIVER_MEMORY}" \
-        "${JAR_DIR}/workload_${CLASS}.jar ${DATA_SIZE}"
+        "${WORKLOAD_JAR_DIR}/workload_${CLASS}.jar ${DATA_SIZE}"
     ;;
 esac
